backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2
python-multipart>=0.0.6
cryptography>=41.0.7
aiosqlite>=0.19.0
slowapi>=0.1.9
httpx>=0.28.0
python-dotenv>=1.0.0
